ci: run coverage on prs (#4738)
Description
---
Run coverage on all PRs

Motivation and Context
---
Currently, this is only run after merge. This is often missed. With fewer PRs, and a more stable code base, I think we can run it on every PR

How Has This Been Tested?
---
Since this is CI related, I'll be using the PR itself to test