fix: computation of vn mmr (#4772)
Description
---
Fix the VNs mmr. The output vector was in random order that resulted in the mmr being kind of random as well.
So the miner and base node had different order of VNs and the blocks were invalidated. More VNs = less probability of passing.

How Has This Been Tested?
---
I manually registered 27 VNs and mined 100 blocks.