fix(dht): zeroize AEAD keys on drop (#4843)
Description
---
Zeroizes authenticated encryption keys (via the `AuthenticatedCipherKey` struct) on drop. Fixes [issue 4842](https://github.com/tari-project/tari/issues/4842).

Motivation and Context
---
Authenticated encryption (AEAD) keys are intended to be zeroized on drop, but the relevant macros are not applied. This work adds the macros.

How Has This Been Tested?
---
Manually tested that `Zeroize` was not supported previously, and is supported now.