fix: delete orphans if they exist (#4868)
Description
---
Orphans should be deleted from the orphan database if they exist. 

Motivation and Context
---
If the block is added to the main chain, it should not exist in the orphan database as well. It will also remove the orphan tips if it removes the block from the pool. 

How Has This Been Tested?
---
Unit test and manually running

fixes: #4867