fix: force wallet sqlite to do checkpoint after db decryption (#4905)
Description
---
Forces the database to do a checkpoint after the last step of the database decryption. 

Motivation and Context
---
The mobile applications decrypt the SQLite wallet database, after which they copy it to a safe location to restore later.  
If we run database decrypt, the actual database file might not yet have the changes applied to it, this forces a checkpoint which will ensure that the changes are committed to on file

How Has This Been Tested?
---
manual

Fixes: #4904