feat: remove tracing_subscriber (#4906)
Description
---
Remove tracing_subscriber and opentelemetry. 

Motivation and Context
---
Fixes #4651 
To my knowledge, this is not used anymore. There seems to be an update to this crate which solves the RUSTSEC warning, but the public API has changed, and it feels better to remove it, and then add it back if it is used or needed in future.

How Has This Been Tested?
---
Cargo build. The use of this feature was opt-in via the `--enable-tracing` arg, so was not run by default.