feat!: add tari address for wallet to use (#4881)
Description
---
This changes the address used by the wallet to not just use the pubkey of another wallet, to a new custom address scheme.

Changes:
- Includes new address scheme : [version][network][pub_key]
- Replaces wallet pubkey with new address all over
- Calculate what address should be is message is received from network public key. 
- Cleans up database to remove broken fields and defunct tables

Motivation and Context
---
We need the wallets to be able to verify that the addresses used by the wallets are for the correct network and are wallets. In interactive transactions, transactions using the wrong pubkeys, will result in transactions eventually being self-canceled by the wallet as they are stale. But one-sided and stealth these will be sent and never recoverable. 

Using the new scheme wallets can verify addresses are on the correct network, and because base_nodes still use pubkeys the addresses differ between base_nodes and wallets. 

How Has This Been Tested?
---
All unit tests passed


Fixes: #4888 
Fixes: #4886 


BREAKING: This changes how the wallet accepts and presents addresses