feat: remove extra validation (#4981)
Description
---
This removes extra validation from transaction protocols. 

Motivation and Context
---
The transaction protocols do a lot of extra unnecessary validation. 
For example:
The code will run add_output(output), which will validate the range_proof and signatures. 
The next line will run tx.build(), which will again validate the range_proof and signatures. 
While these tests are not expensive, they are not free either. 

The second problem is that when we do aggregate transactions such as with branch: `feature-m-of-n` these validations fail as the signatures are partial until right before we build the final transaction

How Has This Been Tested?
---
Passes all unit tests

Fixes: https://github.com/tari-project/tari/issues/4813