fix: remove optionals from wallet set up (#4984)
Description
---
Currently, it is possible to set up a wallet (in creation/recovery/existing mode) without a password. This is not desirable, as in that case, sensitive data might be potentially stored in the database and leak in memory. We address this issue by enforcing passphrase creation.

Motivation and Context
---
For more details see issue #4977.

How Has This Been Tested?
---