feat: add utxo import and export for ffi (#4999)
Description
---
Add the ability to export and import utxos to the faucet

Motivation and Context
---
We need to be able to backup the utxo's without exposing all private information from the wallet. 
This only exposes the limited selection of private keys that are exported

How Has This Been Tested?
---
Unit tests. 

Fixes: https://github.com/tari-project/tari/issues/4954

Open question:
Do we only export the unspent set, or do we export the spent and unspent?