fix: coinbase extra info is not checked (#4995)
This was initially a PR to change the field name of `metadata` in output features to `coinbase_extra`.

This is to reduce confusion, since at least one other field is labelled metadata and looking for this in the code is very confusing with this ambiguity.

However, in making the change, two tests were added:
* Check that *transactions* with coinbase_extra set are rejected as invalid. This turned out not to be the case, because output features are _only checked on orphan block validations_. This is a bug, because invalid transactions (from a stateless perspective) must be rejected as transactions, and not after they have been placed in a block.

* Check that the size of the coinbase_extra field is enforced. This does not appear to be done correctly, since adding a test in the general blockchain end-to-end test did not throw a block with an oversized extra field out.

This then uncovered a very odd definition in the tx weight calculation, which has been marked for clarification.

At this point, this PR is more of an issue, with breaking tests and comments. I'm looking for some clarification on those before making further changes.