docs: multi-party range proof documentation (#5028)
Description
---
Adds documentation for multi-party Bulletproofs+ range proving.

Motivation and Context
---
It is possible for a set of mutually non-trusting players to collaboratively produce an aggregated Bulletproofs+ range proof on commitments for which each player holds a share of the mask for each commitment. This technical note describes a method for producing such a proof.

How Has This Been Tested?
---
Built and viewed locally.