fix: improved encryption key handling (#5027)
Description
---
Improves handling of database encryption keys.

Motivation and Context
---
A [recent PR](https://github.com/tari-project/tari/pull/4984) hardens the codebase's handling of encrypted database fields. It stores the derived key used for encryption as a `Zeroizing` array.

This work changes the key type to be a `Hidden` wrapper of a `SafeArray`, which prevents unintended output of the key and tries to prevent copies and moves.

How Has This Been Tested?
---
Existing tests pass.