fix: console wallet spacing and naming (#5025)
Description
---
Fixes the spacing of the TariAddress field in the transactions panel
Fixes the naming of the Tari address from the public key

Motivation and Context
---
These have been wrong since the update to tari address. The naming was missed when updating, and because the tari address is 2 char's longer, the spacing was out. 

How Has This Been Tested?
---
Manual


Fixes: https://github.com/tari-project/tari/issues/5018
Fixes: https://github.com/tari-project/tari/issues/5017
Fixes: https://github.com/tari-project/tari/issues/4939