feat: add to/from json string for unblinded utxo in wallet ffi (#5098)
Description
---
Adds a conversion to create a json string from an unblinded utxo as well as from json string to unblinded utxo. 

Motivation and Context
---
The wallet ffi needs some way to get all the values inside of the unblinded utxo, this add a to_json and from_json methods to be able to create a json with all the values that should be easy to print out to a file. This same json string can be read into on to create an unblinded output.

How Has This Been Tested?
---
unit tests