fix!: prune mode sync (#5124)
Description
---
Fixes the syncing of pruned nodes

Motivation and Context
---
The utxo and kernel keys are stored as LE bytes, but when reading with the order, they need to be BE bytes to return them in the same order as they were written. When we retrieve these for normal block sync, we call `.sort()` on them with the construction of the block so the order does not matter. 
But prune mode streams the utxo and kernels via MMR positions. For this, the order is important and needs to be returned in the same order as requested. 

This broke to do no cucumber tests testing prune mode sync. This should be fixed when enabling pruned mode. 

How Has This Been Tested?
---
Manual and new unit tests. 

Fixes: https://github.com/tari-project/tari/issues/5099

---
BREAKING CHANGE: REQUIRES THAT BASE NODES RESYNC AND RECALCULATE DB KEYS