refactor(validations): unify validators and remove duplicated logic (#5079)
Description
---
The main idea behind this refactor was to create reusable and composable validation components, so that a particular validation function is found in one and only one of the new components. I tried to preserve the existing logic as much as possible. Also, all the validations are now encapsulated inside the `base_layer/core/src/validation` crate.

The core validation components, reused by every other validators are:
* **Aggregated bodies** (`validation/aggregate_body` crate): used directly as well as part of transactions and block validations.
* **Block headers** (`validation/header` crate): only the `HeaderFullValidator` was created, as the existing logic didn't required splitted validators for internal consistency and chain link verification.

Additionally, as a convenience layer on top of aggregated bodies, we have both **transactions** (`validation/transaction` crate) and **block bodies** (`validation/block_body` crate) validators. Some of them add some extra validations on top of the aggregate body ones. Maybe in the future some of these could be pushed down into the aggregate body validations to make them more generic.

To facilitate code reviewing, this is the list with the correspondence between old validators and new ones:
* The `AggregateBody.validate_internal_consistency` method was removed from the struct into a new, separated, reusable `AggregateBodyInternalConsistencyValidator` component. This required to add the `base_node` feature to the `tari_core` dependency of the `wallet_ffi`, so if for whatever reason this is not acceptable (binary size?) we will need to move back the validation to the `AggregateBody` struct.
* `HeaderValidator` -> `HeaderFullValidator`
* `MempoolValidator` was a trait with 3 different implementations, all could be merged into a single one and implemented with `TransactionFullValidator`
* `BodyOnlyValidator` -> `BlockBodyFullValidator`
* `OrphanBlockValidator` -> `BlockBodyInternalConsistencyValidator`
* `BlockHeaderSyncValidator` (`base_layer/core/src/base_node/sync/header_sync/validator.rs`) still exists for API convenience, but delegates the actual validation rules into the new `HeaderFullValidator` (used by the mempool as well), so we avoid duplicities
* `async_validator.rs` is removed, we are now reusing a `BlockBodyFullValidator` wrapped in a `tokio::task`

Motivation and Context
---
Currently there is duplicate validation logic in multiple places. This makes it hard to reason about whether it is correct according to the RFC, but it also means that as new rules are added or fixed, it could be done in one place, but not the other.

Solves https://github.com/tari-project/tari/issues/5042

How Has This Been Tested?
---
Existing unit tests pass, with the exception of: 
* `block_event_and_reorg_event_handling`
* `test_retrieve`
* `test_base_node_wallet_rpc`

I ignored them as it seems that they will need to be rewritten taking into account that the mempool now performs a lot more validations on incoming transactions.