fix(test): broken address test (#5134)
Description
---
Fixes a broken [address test](https://github.com/tari-project/tari/blob/5f7f04b599ab49765466fb7e0faa37e2dfec4afe/base_layer/common_types/src/tari_address.rs#L304-L335) that exhibits inconsistent failure.

Closes [issue 5133](https://github.com/tari-project/tari/issues/5133).

Motivation and Context
---
The test in question claims to detect failures caused by bad network bytes. However, it attempts to do this by mutating the checksum, setting the second address byte to a fixed value that happens to match the checksum mutation, and checking for a checksum failure when the result is decoded using the same network used when generating the address.

It's not clear why the test is structured this way. What it's really doing is simulating the effect of two mutations, which the checksum algorithm is not guaranteed to detect (and which has caused CI failures).

This PR restructures the test to perform two separate but related tests.

In the first test, we generate a valid address on one network and attempt to decode it for another network. The checksum algorithm guarantees this will always be detected. In the second test, we mutate the checksum and attempt to decode it for the same network. The checksum algorithm guarantees this will also always be detected.

How Has This Been Tested?
---
Updated test passes. Manually ran 100000 iterations to ensure no obvious spurious failures occur.