fix: feature flag separation for validation (#5137)
Description
---
Removes OpenSSL from the WalletFFI lib as its not required or used by it. 
Fixes validation feature flags. 

Motivation and Context
---
Recent refactor added `base_node` feature flag to the walletFFI. This should not be done as it adds in all the base node 
specific crates that not required by the wallet like randomX, lmdb etc. 
This splits the validation traits and implementations between the features `transactions` and `base_node`.

How Has This Been Tested?
---
Manual and unit tests

Fixes: https://github.com/tari-project/tari/issues/4974