Update Diesel dependency
libsqlite3 was pinned to the highest version we could run as defined by
the version of diesel we were running. We were running the highest
version of diesel available without making a major version increment. To
update libsqlite we also need to update diesel. It's a slightly larger
undertaking but should be worthwhile long term.